scrapingcourse_scraper/requirements.txt
scrapy
sqlalchemy
psycopg2-binary
Pillow
requests
pytest
coverage
pytest-cov